requirements.txt
django
graphene-django
django-cors-headers
django-graphql-jwt
